api/requirements.txt
click==6.7
Flask==1.0.2
Flask-Cors==3.0.4
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.4
PyMySQL==0.8.1
six==1.11.0
Werkzeug==0.14.1
